Moved `csv` to dev dependencies
