+++ Cargo.toml
@@ -37,7 +37,6 @@ serde = { version = "1.0.114", features = ["derive"] }
 dirs = "3.0.1"
 itertools = "0.9.0"
 ordered-float = "2.0.0"
-csv = "1.1.3"
 reqwest = "0.10.7"
 lazy_static = "1.4.0"
 uuid = { version = "0.8.1", features = ["v4"] }
@@ -46,3 +45,4 @@ thiserror = "1.0.20"
 
 [dev-dependencies]
 anyhow = "1.0.32"
+csv = "1.1.3"
