+++ src/pipelines/sentiment.rs
@@ -58,10 +58,6 @@ use crate::common::error::RustBertError;
 use crate::pipelines::sequence_classification::{
     SequenceClassificationConfig, SequenceClassificationModel,
 };
-use serde::Deserialize;
-use std::error::Error;
-use std::fs;
-use std::path::PathBuf;
 
 #[derive(Debug, PartialEq)]
 /// Enum with the possible sentiment polarities. Note that the pre-trained SST2 model does not include neutral sentiment.
@@ -153,24 +149,4 @@ impl SentimentModel {
         }
         sentiments
     }
-}
-
-#[derive(Debug, Deserialize)]
-struct Record {
-    sentence: String,
-    label: i8,
-}
-
-pub fn ss2_processor(file_path: PathBuf) -> Result<Vec<String>, Box<dyn Error>> {
-    let file = fs::File::open(file_path).expect("unable to open file");
-    let mut csv = csv::ReaderBuilder::new()
-        .has_headers(true)
-        .delimiter(b'\t')
-        .from_reader(file);
-    let mut records = Vec::new();
-    for result in csv.deserialize() {
-        let record: Record = result?;
-        records.push(record.sentence);
-    }
-    Ok(records)
 }
